Moving library example one level down.
