--- .dockerignore
@@ -1,2 +1,0 @@
-/**/*
-!build/libs/GraphQLPlayground*.jar
