+++ examples/library/.dockerignore
@@ -1,0 +1,2 @@
+/**/*
+!build/libs/library.jar
