+++ examples/library/Dockerfile
@@ -1,0 +1,4 @@
+FROM openjdk:11
+COPY build/libs/library.jar /app/app.jar
+WORKDIR /app
+CMD ["java", "-jar", "app.jar"]
