+++ examples/library/build.gradle
