+++ examples/library/src/main/kotlin/com/chaosonic/graphql/library/GraphQLApplication.kt
@@ -1,4 +1,4 @@
-package com.chaosonic.graphql.example
+package com.chaosonic.graphql.library
 
 import org.springframework.boot.SpringApplication
 import org.springframework.boot.autoconfigure.SpringBootApplication
