+++ examples/library/src/main/resources/application.yml
