+++ examples/library/src/main/resources/data/data.yml
