+++ examples/library/src/main/resources/graphql/.graphqlconfig
