+++ examples/library/src/main/resources/graphql/schema.graphqls
