+++ examples/library/src/test/resources/test-schema.graphqls
